# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/actor-add": {
            "post": {
                "description": "Adds a new actor with the given details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Adds a new actor",
                "parameters": [
                    {
                        "description": "Actor to add",
                        "name": "actor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Actor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully added the actor",
                        "schema": {
                            "$ref": "#/definitions/models.Actor"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.Actor"
                        }
                    },
                    "500": {
                        "description": "Error creating actor",
                        "schema": {
                            "$ref": "#/definitions/models.Actor"
                        }
                    }
                }
            }
        },
        "/v1/actor-delete/{id}": {
            "delete": {
                "description": "Deletes the actor with the specified ID, including removing all associated movies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Deletes an actor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Actor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted the actor",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid actor ID or URL format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Actor not found or could not be deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/actor-edit/{id}": {
            "put": {
                "description": "Edits an actor with the specified ID based on the given update fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Edits an existing actor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Actor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated the actor",
                        "schema": {
                            "$ref": "#/definitions/models.Actor"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or actor ID",
                        "schema": {
                            "$ref": "#/definitions/models.Actor"
                        }
                    },
                    "404": {
                        "description": "Actor not found",
                        "schema": {
                            "$ref": "#/definitions/models.Actor"
                        }
                    },
                    "500": {
                        "description": "Failed to save actor",
                        "schema": {
                            "$ref": "#/definitions/models.Actor"
                        }
                    }
                }
            }
        },
        "/v1/actor-list": {
            "get": {
                "description": "Retrieves a list of all actors, including their associated movies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Lists all actors",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all actors",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Actor"
                            }
                        }
                    },
                    "500": {
                        "description": "Error retrieving actors",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Actor"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Actor": {
            "type": "object",
            "properties": {
                "dateOfBirth": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "movies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Movie"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Movie": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Actor"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "releaseDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}